# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Avora",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Create Assessment",
        "operationId": "create_assessment_api_v1_assessments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "List Assessments",
        "operationId": "list_assessments_api_v1_assessments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssessmentOut"
                  },
                  "title": "Response List Assessments Api V1 Assessments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/assessments": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "My Assessments",
        "operationId": "my_assessments_api_v1_me_assessments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyAssessmentScore"
                  },
                  "type": "array",
                  "title": "Response My Assessments Api V1 Me Assessments Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/observations": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Create Observation",
        "operationId": "create_observation_api_v1_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/students/{student_id}/seed-readiness": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Seed Readiness",
        "description": "A tutor's own starting assessment of a student, per topic.\n\nThis is step 5 of the cold start (spec \u00a77.1), and it deliberately cannot\nhappen earlier: students attach themselves by invite code, so a tutor\nfinishing setup has no students to rate. It is also optional, because for\nmany classes it never happens \u2014 the product has to work when it does not.\n\nWhat is written is `Evidence` with `source_type=tutor_estimate`, which means\nit goes through exactly the same pipeline as every other source and needs no\nparallel code path (PROD-9's shape, applied to a new source). It is\nself-declared and labelled as such wherever it is shown (PROD-8), and it\nloses weight against real marked work as that arrives (\u00a77.3).\n\nRe-seeding the same topic replaces the previous estimate rather than adding\na second one: this is the tutor's current opinion, not a history of their\nopinions, and stacking them would let one topic be seeded into a score.\n\n**Known gap (RISK-5).** The estimate's numeric value drives the v1 engine.\nReadiness Engine v2 \u2014 the default served engine \u2014 computes topic mastery from\nmarked questions and reads Evidence only as a \"practised\" flag, so it does\nnot consume the estimate's percentage. This is not unique to seeding: every\nEvidence-only source (tutor observations, entered mocks) has the same limit\nunder v2, and closing it means teaching v2 to score from Evidence, which is\nthe RISK-5 convergence work, not a cold-start change. Until then the seed's\nnumber is fully honoured only where v1 answers.",
        "operationId": "seed_readiness_api_v1_students__student_id__seed_readiness_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedReadinessIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Seed Readiness Api V1 Students  Student Id  Seed Readiness Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/students/{student_id}/observations": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "List Observations",
        "operationId": "list_observations_api_v1_students__student_id__observations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObservationOut"
                  },
                  "title": "Response List Observations Api V1 Students  Student Id  Observations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentCreate": {
        "properties": {
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "type": {
            "type": "string",
            "pattern": "^(mock|test)$",
            "title": "Type",
            "default": "mock"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/AssessmentScoreIn"
            },
            "type": "array",
            "title": "Scores",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "title",
          "date"
        ],
        "title": "AssessmentCreate"
      },
      "AssessmentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "score_count": {
            "type": "integer",
            "title": "Score Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_id",
          "title",
          "type",
          "date",
          "score_count"
        ],
        "title": "AssessmentOut"
      },
      "AssessmentScoreIn": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "marks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Marks"
          },
          "max_marks": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Marks"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "marks",
          "max_marks"
        ],
        "title": "AssessmentScoreIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyAssessmentScore": {
        "properties": {
          "assessment_id": {
            "type": "integer",
            "title": "Assessment Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "subject_id": {
            "type": "integer",
            "title": "Subject Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "topic_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Title"
          },
          "marks": {
            "type": "integer",
            "title": "Marks"
          },
          "max_marks": {
            "type": "integer",
            "title": "Max Marks"
          },
          "pct": {
            "type": "number",
            "title": "Pct"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "title",
          "type",
          "date",
          "subject_id",
          "subject_name",
          "topic_id",
          "topic_title",
          "marks",
          "max_marks",
          "pct"
        ],
        "title": "MyAssessmentScore"
      },
...
      "ObservationCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "comment": {
            "type": "string",
            "minLength": 1,
            "title": "Comment"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "comment"
        ],
        "title": "ObservationCreate"
      },
      "ObservationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "topic_id",
          "comment",
          "rating",
          "created_at"
        ],
        "title": "ObservationOut"
      },
...
      "SeedReadinessIn": {
        "properties": {
          "topics": {
            "items": {
              "$ref": "#/components/schemas/SeedTopicScore"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Topics"
          }
        },
        "type": "object",
        "required": [
          "topics"
        ],
        "title": "SeedReadinessIn"
      },
      "SeedTopicScore": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id"
          },
          "score_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score Pct"
          }
        },
        "type": "object",
        "required": [
          "topic_id",
          "score_pct"
        ],
        "title": "SeedTopicScore"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}